feat(typegen): nicer generated package name
This can impact the end-user when they, for example, are using the
`types` TS compiler option. In that situation, expressing a whitelist of
TypeScript type-packages, in a Nexus app, it would be important that the
user also include typegen or else TSC would stop seeing it (forcing the
user to provide custom typegen config). To make it easier for users to
do that, let them refer to a pretty name, rather than something hard to
predict, remember, and looks ugly.