feat(typegen): nicer generated package name (#248)
* feat(typegen): nicer generated package name

This can impact the end-user when they, for example, are using the `types` TS compiler option. In that situation, expressing a whitelist of TypeScript type-packages, in a Nexus app, it would be important that the user also include typegen or else tsc would stop seeing it (forcing the
user to do Nexus typegen config). To make it easier for users to do that, this change allows them refer to a pretty name, rather than something hard to predict, remember, and ugly looking.