feat: allow control over nexus schema import id (#408)
Currently the typegen output is designed with the assumption that @nexus/schema will be available at node_modules/@nexus/schema. This assumes that either the project depends on @nexus/schema or that @nexus/schema has been hoisted. The former is fine but does not support being wrapped by another tool, while the latter is liable to result in module not found errors for hard-to-debug reasons (especially users who know little or nothing about the internals).