Update 04-chapter-3-adding-mutations-to-your-api.mdx
This is something I came across as I was going through the tutorial. Not explicitely setting an export key on contextType appears to throw a Typescript error:

`Property 'export' is missing in type '{ module: string; alias: string; }' but required in type 'TypingImport'.`

Removing the alias key also appears to suppress the TypeScript error.